add Docker image for easy running
Even though there's not much to this project in terms of getting
started, Docker helps ensure isolation and repeatability, and the
requirements here are very minimal. This adds a Dockerfile and
instructions for how to use it. It also beefs up the instructions in
general, based on what I've learned while running this thing.