make "example" properties create more load
When I first came to this project, I assumed the example properties
file was set to create some decent load. It isn't. Through running it,
reading the code, and examining Blueflood performance charts that
result from running it, I realized that the existing settings are
extremely low in terms of load. This updates them to be more like what
you'd expect from a "performance" or "load test" project. Instead of
generating tens of data points for a few tenants, now it'll generate
hundreds of thosands of them for a hundred tenants.