generate metrics for more than one tenant
As originally written, this project would generate metrics for
multiple tenants, picking them at random. Some time more recently,
someone added the ability to authenticate a real user and use that
user's tenant id. In doing so, they made it so only one tenant id was
used *always*, even if not using a real user. This brings back the
code that randomizes the tenant id when *not* using a real user.