query for 1 minute of annotations instead of 1 day
When I ramped the load test up on staging, it started making query
nodes run out of memory. I traced it back to the annotations queries
here, which appears to run a query similar to the others, but it fails
to account for the fact that "annotations" are just little json blobs
to keep track of discrete events. They don't form any time series and
aren't rolled up. It's fine to query for a day's metrics because
you'll get downsampled data, but annotations just accumulate, and if
you run the load test for the whole day, by the end of it, there are
far too many annotations to fit into memory when generating the
payload.