add flag to require non-empty queries
While testing with various permutations, it's nice to know that
queries are actually returning some data and not just succeeding with
no results. At the same time, depending on your configuration, that
might cause more problems than it solves, so this makes a flag to turn
that behavior on or off.