remove "even tenants have delays" logic
The project formerly introduced random delayed metrics for tenants
with an even tenant id. When you authenticate as a normal user and
turn of the /ingest/multi endpoint, you're producing metrics for only
one tenant: yours. That means the possibility of delay can't be based
on tenant id. This removes that condition.