increase possible metric name permutations
This project was written to be able to generate metric names that were
randomized in only one token of the name, which limits its usefulness
in testing metric tokens at scale because all the tokens but one are
the same.

In other words, even with a long name pattern like
org.example.a.1.b.2.c.3.d.4.e.5.f.6.%d and your randomization factor
set to 1000, you still only get 1001 tokens because all tokens leading
up to the final one never change. By interspersing more randomness in
that name, the number of tokens rises dramatically.

This adds the ability to do that through a new property, keeping the
original behavior because tests rely on it extensively.